notebooks/pyproject.toml
[project]
name = "nb"
version = "0.1.0"
requires-python = ">=3.12, <4.0"
dependencies = [
    "marimo[recommended]>=0.18.0",
    "matplotlib>=3.10.6",
    "plotly>=6.5.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torchmetrics>=1.8.2",
    "torchvista>=0.2.6",
    "torchviz>=0.0.3",
    "transformers[retrieval,timm,torch-vision]>=4.57.3",
    "wigglystuff>=0.2.8",
]
